Keep logspace functions together in resample
